refactor: extract chat session workspace
